requirements.txt
opencv-python
sentence_transformers
Pillow
pymongo